// types/guru.ts
export interface GuruProfile {
  id: number;
  nama: string;
  nuptk: string | null;
  jk: string;
  tempat_lahir: string | null;
  tanggal_lahir: string | null;
  nik: string | null;
  nip: string | null;
  jenis_ptk: string | null;
  gelar: string | null;
  jenjang: string | null;
  prodi: string | null;
  sertifikasi: string | null;
  tmt_kerja: string | null;
  tugas_tambahan: string | null;
  kompetensi: string | null;
  user_id: number | null;
}